# backend/API-docs.yaml
...
openapi: 3.0.1
info:
  title: Maximum Visibility Dive API
  description: The Maximum Visibility API is a recteational-technical dive planner, it can generate no-decompression dive tables or a decompression plan and gas consumption calculation per input in metric measurements. As an extension it can save past dives in a dive log and save the adjustable gas and conservativism settings in a user profile.
  version: "1.0"
servers:
  - url: https://max-vis-app-umuud.ondigitalocean.app/api
security:
  - bearerAuth: []
...
paths:
...
  /settings:
    get:
      tags:
        - settings
      summary: Returns user settings.
      description: Returns user settings.
      responses:
        "200":
          description: Success.
      servers:
        - url: https://max-vis-app-umuud.ondigitalocean.app/api
    patch:
      tags:
        - settings
      summary: Updates user settings.
      description: Updates user settings.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/settings_body"
            examples:
              "0":
                value: |-
                  {
                          "ascentSpeed": 10,
                          "descentSpeed": 20,
                          "gradFLow": 0.4,
                          "gradFHigh": 0.7,
                          "tissues": [
                              {
                                  "halfTime": 4,
                                  "mValue": 3.24
                              },
                              {
                                  "halfTime": 8,
                                  "mValue": 2.44
                              },
                              {
                                  "halfTime": 15,
                                  "mValue": 2.1
                              },
                              {
                                  "halfTime": 30,
                                  "mValue": 1.8
                              },
                              {
                                  "halfTime": 60,
                                  "mValue": 1.57
                              },
                              {
                                  "halfTime": 120,
                                  "mValue": 1.45
                              },
                              {
                                  "halfTime": 240,
                                  "mValue": 1.37
                              },
                              {
                                  "halfTime": 500,
                                  "mValue": 1.39
                              }
                          ],
                          "gasMixes": [
                              {
                                  "o2": 0.21,
                                  "he2": 0,
                                  "n2": 0.79,
                                  "diveMod": 56.6,
                                  "decoMod": 66.1,
                                  "_id": "62c04086b6e457e2eafb02dd"
                              }
                          ]
                  }
      responses:
        "200":
          description: Success.
      servers:
        - url: https://max-vis-app-umuud.ondigitalocean.app/api
    servers:
      - url: https://max-vis-app-umuud.ondigitalocean.app/api
  /settings/gasmix:
    post:
      tags:
        - settings
      summary: Creates new breathing gas mix.
      description: Creates new breathing gas mix.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/settings_gasmix_body"
            examples:
              "0":
                value: "\n            {\n                \"o2\": 0.28,\n                \"he2\": 0\n            }\n "
      responses:
        "200":
          description: Success
      servers:
        - url: https://max-vis-app-umuud.ondigitalocean.app/api
    servers:
      - url: https://max-vis-app-umuud.ondigitalocean.app/api
  /settings/gasmix/{o2}:
    delete:
      tags:
        - settings
      summary: Returns user settings.
      description: Returns user settings.
      parameters:
        - name: o2
          in: path
          description: Gasmix O2 content.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: "0.21"
        - name: he2
          in: query
          description: Gasmix He2 content.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: "0.17"
      responses:
        "200":
          description: Success.
      servers:
        - url: https://max-vis-app-umuud.ondigitalocean.app/api
    servers:
      - url: https://max-vis-app-umuud.ondigitalocean.app/api
components:
  schemas:
...
    settings_tissues:
      type: object
      properties:
        halfTime:
          type: integer
        mValue:
          type: number
      example: null
...
    settings_body:
      type: object
      properties:
        gasMixes:
          type: array
          items:
            $ref: "#/components/schemas/settings_gasMixes"
        tissues:
          type: array
          items:
            $ref: "#/components/schemas/settings_tissues"
        ascentSpeed:
          type: integer
        gradFHigh:
          type: number
        gradFLow:
          type: number
        descentSpeed:
          type: integer
    settings_gasMixes:
      type: object
      properties:
        o2:
          type: number
        he2:
          type: integer
        n2:
          type: number
        diveMod:
          type: number
        decoMod:
          type: number
        _id:
          type: string
      example: null
...
    settings_gasmix_body:
      type: object
      properties:
        o2:
          type: number
        he2:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT